Split vp8 decoder into modules
